+++ src/main/authentication/authentication.ts
@@ -7,18 +7,14 @@ import {
   AUTH_RETURN_URL,
   MEDIA_VERIFICATION_RETURN_URL,
   ADMIN_AUTH_RETURN_URL,
-  FRONTEND_URL,
-  CFT_IDAM_URL,
 } from '../helpers/envUrls';
-import axios from 'axios';
-import jwt_decode from 'jwt-decode';
+import {cftIdamAuthentication} from "./cftIdamAuthentication";
 
 const AzureOIDCStrategy = require('passport-azure-ad').OIDCStrategy;
 const passport = require('passport');
 const authenticationConfig = require('./authentication-config.json');
 const logger = Logger.getLogger('authentication');
 const CustomStrategy = passportCustom.Strategy;
-const querystring = require('querystring');
 const accountManagementRequests = new AccountManagementRequests();
 
 /**
@@ -30,7 +26,6 @@ function oidcSetup(): void {
   let identityMetadata;
   let adminIdentityMetadata;
   let mediaVerificationIdentityMetadata;
-  let cftIdamClientSecret;
 
   if(process.env.CLIENT_SECRET) {
     clientSecret = process.env.CLIENT_SECRET;
@@ -60,12 +55,6 @@ function oidcSetup(): void {
     mediaVerificationIdentityMetadata = process.env.MEDIA_VERIFICATION_CONFIG_ENDPOINT;
   } else {
     mediaVerificationIdentityMetadata = config.get('secrets.pip-ss-kv.MEDIA_VERIFICATION_CONFIG_ENDPOINT') as string;
-  }
-
-  if(process.env.CFT_IDAM_CLIENT_SECRET) {
-    cftIdamClientSecret = process.env.CFT_IDAM_CLIENT_SECRET;
-  } else {
-    cftIdamClientSecret = config.get('secrets.pip-ss-kv.CFT_IDAM_CLIENT_SECRET') as string;
   }
 
   logger.info('secret', clientSecret ? clientSecret.substring(0,5) : 'client secret not set!' );
@@ -150,39 +139,7 @@ function oidcSetup(): void {
   piAadVerifyFunction,
   ));
 
-  passport.use('cft-idam', new CustomStrategy(
-    async function(req, callback) {
-
-      const params = {
-        client_id: 'app-pip-frontend',
-        client_secret: cftIdamClientSecret,
-        grant_type: 'authorization_code',
-        redirect_uri: FRONTEND_URL + '/cft-login/return',
-        code: req.query.code as string,
-      };
-
-      const tokenRequest = axios.create({baseURL: CFT_IDAM_URL, timeout: 10000});
-
-      try {
-        const response = await tokenRequest.post('/o/token', querystring.stringify(params), {
-          headers: {
-            'Accept': 'application/json',
-            'Content-Type': 'application/x-www-form-urlencoded',
-          },
-        });
-
-        const data = response.data;
-        const jwtToken = jwt_decode(data.id_token);
-        jwtToken['flow'] = 'CFT';
-
-        callback(null, jwtToken);
-
-      } catch (e) {
-        console.log(e);
-      }
-    },
-  ));
-
+  passport.use('cft-idam', new CustomStrategy(cftIdamAuthentication));
 }
 
 /**
